rafthttp: move test-only functions to '_test.go'
Not used in actual code base, only used in tests
